feat: add initializer config
