+++ Procfile
@@ -1,2 +1,3 @@
 app: sh -c "cd ./web/app && yarn dev"
+core: sh -c "cd ./services/core && bin/rails s"
 tasks: sh -c "cd ./services/tasks/ && exec air -c .air.toml"
