+++ services/core/Gemfile
@@ -6,13 +6,13 @@ git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 ruby "3.0.3"
 
 gem "bootsnap", require: false
+gem "devise"
 gem "jbuilder"
 gem "jwt"
 gem "pg", "~> 1.1"
 gem "puma", "~> 5.0"
 gem "rails", "~> 7.0.3", ">= 7.0.3.1"
 gem "tzinfo-data", platforms: %i[mingw mswin x64_mingw jruby]
-gem "devise"
 
 # Use Redis adapter to run Action Cable in production
 # gem "redis", "~> 4.0"
