+++ services/core/app/controllers/application_controller.rb
@@ -7,7 +7,7 @@ class ApplicationController < ActionController::Base
 
   respond_to :json
 
-  # before_action :underscore_params!
+  before_action :underscore_params!
   before_action :configure_permitted_parameters, if: :devise_controller?
   before_action :authenticate_user
 
@@ -18,19 +18,17 @@ class ApplicationController < ActionController::Base
   end
 
   def authenticate_user
-    if request.headers["Authorization"].present?
-      authenticate_or_request_with_http_token do |token|
-        begin
-          jwt_payload = JWT.decode(token, Rails.application.secrets.secret_key_base).first
-          @current_user_id = jwt_payload["id"]
-        rescue JWT::ExpiredSignature, JWT::VerificationError, JWT::DecodeError
-          head :unauthorized
-        end
-      end
+    return if request.headers["Authorization"].blank?
+
+    authenticate_or_request_with_http_token do |token|
+      jwt_payload = JWT.decode(token, Rails.application.secrets.secret_key_base).first
+      @current_user_id = jwt_payload["id"]
+    rescue JWT::ExpiredSignature, JWT::VerificationError, JWT::DecodeError
+      head :unauthorized
     end
   end
 
-  def authenticate_user!(options = {})
+  def authenticate_user!(_options = {})
     head :unauthorized unless signed_in?
   end
 
@@ -40,5 +38,9 @@ class ApplicationController < ActionController::Base
 
   def signed_in?
     @current_user_id.present?
+  end
+
+  def underscore_params!
+    params.deep_transform_keys!(&:underscore)
   end
 end
