+++ services/core/app/controllers/sessions_controller.rb
@@ -1,8 +1,8 @@
 class SessionsController < Devise::SessionsController
   def create
-    user = User.find_by_email(sign_in_params[:email])
+    user = User.find_by(email: sign_in_params[:email])
 
-    if user && user.valid_password?(sign_in_params[:password])
+    if user&.valid_password?(sign_in_params[:password])
       @current_user = user
     else
       render(
