+++ services/core/app/views/users/_user.json.jbuilder
@@ -1,2 +1,2 @@
-json.(user, :id, :email, :username, :image)
+json.call(user, :id, :email, :username, :image)
 json.token user.generate_jwt
