+++ services/core/config/initializers/jbuilder.rb
@@ -1,0 +1,1 @@
+Jbuilder.key_format camelize: :lower
