+++ services/core/spec/models/user_spec.rb
@@ -1,4 +1,4 @@
-require 'rails_helper'
+require "rails_helper"
 
 RSpec.describe User, type: :model do
   pending "add some examples to (or delete) #{__FILE__}"
